Case Watch
Adds a case watcher feature - similar to flagging, but it's stored locally and only you can see it. Also supports notes. Use - go the view case page, click the flag button, then click "watch this case". Notes are optional, and can be put in the text box next to "watch this case". To view your watched cases, click the "Case Watch" button in the top left of your screen.